adopt new material-structure

+++ merlict_c89/mliMaterials.h
@@ -12,27 +12,27 @@
 #include "mliName.h"
 
 struct mliMaterialsCapacity {
-        uint32_t num_surfaces;
-        uint32_t num_media;
-        uint32_t num_boundary_layers;
+        uint64_t num_surfaces;
+        uint64_t num_media;
+        uint64_t num_boundary_layers;
 };
 
 struct mliMaterialsCapacity mliMaterialsCapacity_init(void);
 
 struct mliMaterials {
-        uint32_t num_surfaces;
+        uint64_t num_surfaces;
         struct mliSurface *surfaces;
         struct mliName *surface_names;
 
-        uint32_t num_media;
+        uint64_t num_media;
         struct mliMedium *media;
         struct mliName *medium_names;
 
-        uint32_t num_boundary_layers;
+        uint64_t num_boundary_layers;
         struct mliBoundaryLayer *boundary_layers;
         struct mliName *boundary_layer_names;
 
-        uint32_t default_medium;
+        uint64_t default_medium;
 };
 
 int mliMaterials_malloc(
